gateway/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
PyJWT
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.1.0
aio-pika
pika==1.3.2
httpx
requests

